Throw in Token.create
As of https://github.com/aptible/aptible-auth-ruby/pull/49, Token
behaves like other HyperResource, and doesn't throw on .create (which is
probably not ideal, but is consistent).

Unfortunately, some clients relie on create throwing (namely:
`aptible-cli`), and those clients don't pin a specific version, so we
have no choice but to preserve backwards compatibility here.

Those clients also rely on the error being a OAuth2 error, so there's
that too.