fix up ci.yml
we were setting an older bundler version in the GHA, but that was not getting used during testing of ruby 3.3 and 3.4. Those jobs would have failed, if it did since bundle 1.17 is not compatible (and errors) with ruby 3.4.

The jobs were not failing, because they were using the "BUNDLED WITH" version at the bottom of the Gemfile.lock. That is version 2.7.2.

Remove the bundler-version from the GHA, to avoid any confusion.

Also: use the ruby/setup-ruby action to run bundle install. https://github.com/ruby/setup-ruby/tree/v1.269.0/?tab=readme-ov-file#caching-bundle-install-automatically